Changed, added, removed some text
Changed any text referring to 'Second in Charge' to either 'Administrator' or 'the Administrator'.
Changed examples showing 'ARTCvan#1234' to 'ARTCvan#4974'.
Changed section 'Special Formatting' to use h2 instead of h1.
Changed some text that said 'Example would be' to 'An example would be' (depending if the line referred to multiple cmds)

Removed Level roles from the role requirements table, may be added later.

Added dot points to the "How do you moderate" section to streamline the look so that descriptions are in plain text, whilst titles are using h2, examples are quotes.
Added text to state that using other Discord clients may result in an account ban by Discord, so it is recommended to stick with Discord official client. 
Added text to state that this will be replaced by a Wiki or Pages in GitHub soon.